Audio sharing bug fixed
Turning off video sharing now does not affect audio sharing.

in the function handleCheckAudio,
when selecting an audio sender, it checks the condition if
track.kind  is audio  Or sender.dtmf is not null as I assume that video senders do not have 'dtmf' property.

On the other hand, in handleCheckVideo,
it will select one of the senders if they have track.kind is video Or sender.dtmf is null.

When turning off video sharing,
it previously set the HTMLMediaElement to null, therefore, it does not process any audio/video tracks.
Now, new mediaStream object made with audio tracks is placed instead.